Add UART driver for nrf52
This patch makes the nrf52 able to output some things on the UART that
is connected to the stm32.  Since blackmagic on the stm32 is configured
to act as an USB-UART device, we are able to see the output using a
terminal emulator on the dev machine.  For reference, this is what I
used to test:

    $ picocom --imap lfcrlf  --baud 38400  /dev/ttyACM1

I used 38400 baud, because that is the default of the usbuart module of
blackmagic.  But we can crank it up once we see that it works well.

I added some traces in the buttons code, as an example that doesn't send
too much data.