Fix case of gfx functions
There is a mismatch of case (snake_case vs camelCase) for some
functions.  It is not only cosmectic, some functions call
ssd1306_drawPixel, while the real function is called ssd1306_draw_pixel.
My guess is that the functions that use the wrong name are not actually
used.  So the compilation passes (giving some implicit declaration
warnings), and those callers are garbage collected during linking, so
there is no undefined reference.

It's better to just fix the warnings and use the right names.