Introduce external flash driver, flash mode, flash client
Introduce basic functions to access the external flash
(drivers/flash.{c,h}).  For now, I chose to add some restrictions, such
as making reads and writes operator on 128 bytes, for simplicity.  Also,
writes must be aligned on multiples of 128 bytes.  The reason for that
last one is that it's not possible to write across two pages of flash
(on page being 256 bytes), the write just wraps around to the beginning
of the first page.  So if we let the callers write at any
address, they will get some unexpected results.

Then, this patch adds a "flash mode".  The intent is that we can put the
badge in this mode to program (erase/write/read) the flash from PC.
When we are done, we reset the badge.  In this mode, the badge accepts
and response to text commands through the UART.

Finally, this patch adds a Python client that talks to the badge when it
is in this mode.