Add .clang-format file
Based on the default style (LLVM), but with 4 spaces instead.

To format your file in-place, use "clang-format -i <file>".  When
touching files that don't comply with the coding style yet, this can
introduce many unrelated changes and just add noise to the patch.

Instead, it's possible to use git-clang-format as well, to format only
the lines that were touched by a specific commit.