Add support for images on external flash
This patch adds support for putting bitmaps on the external flash and
rendering them (although quite slowly).  This includes:

- A tool to concatenate files into one and produce a table of contents
in the for of a C header file (that's pack_flash.py).
- Modifications to gen_image.py to say that an image should be placed on
the external flash.  When the --external-flash flag is passed, the script
will not place the bytes in the generated C source code, but in a
separate file (which will end up packed by pack_flash.py).  Instead of
pointing to the bytes directly, the bitmap_ext structure (also
introduced in this commit) points to the entry of the table of contents
generated by pack_flash.py.
- Modifications to flash_client.py to write a file passed as argument to
the flash, and a Makefile target to send the packed file.
- A function that takes a bitmap_ext structure and draws it to the
screen.

As a test, I made the "flash mode" display an image which comes from the
external flash.  When first accessing the flash mode, the image is of
course not in the flash yet, so the screen should display white (value
of flash bytes when it's erased) or whatever else is currently in the
flash.

But if you send the packed file (make flash-external-flash) and re-enter
flash mode, you should see the image.