Migrating interface to new scheme
There is now a difference between home and main menu, the home menu is the front page and
the main menu is the equivalent of what we had before. We use the occasion to move the main
menu stuff to a specific file and clean up the main file.

The menu system was hardcoded in black and white. With the color display it is usefull to
configure the text and background color.

Adding menu_handler_init in the init sequence. Before the first menu_init was called by
the main init and was not triggered by a user input. If the first menu_init was called
by a user input a input trigger will follow. By registering the menu to the controls handler
only one time this is avoided and we dont need to register for every menu_init anyway.

Adding parameter to show_home_menu to set the state of the front page when redrawing it

removing app_glue.h