Add CLI over UART
This patch adds CLI functionality over UART (over USB) using Nordic's
CLI module.

The UART is now completely managed by the CLI, we should not try to read
or write from it directly.  The existing uart support is kept for the
flash mode, but it should not be used anywhere else.

It usually does not work right after flashing the program.  Power cycle
the badge, then use:

  picocom --baud 115200 /dev/ttyACM1