Make space for config area in external flash
When flashing, we only erase the area we need (with a granularity of 4
kiB, the smallest that the flash allows).

We also make sure that the data we are writing does not overwrite the
last 4 kiB block of flash, reserved for persistent config.

In practice, this patch changes the "erase" command of the flash mode to
take an address as a parameter, and makes it erase the 4 kiB block that
contains this address.  There is no more command to erase the whole
chip, but it could always be added back later.