Fix bug in nrf_ringbuf, fix UART hang
There is a bug in the nrf_ringbuf implementation in the version of the
SDK we use.  nrf_ringbuf_alloc sometimes miscalculates the amount of
available buffer, which leads to a UART hang.

1. rx_try tries to allocate 255 bytes using nrf_ringbuf_alloc.
2. nrf_ringbuf_alloc returns a length of 255 bytes, even though less is
available.
3. the following nrf_ringbuf_put fails, because we pass it a length of
255 bytes, but it determines (correctly) that this is more than the
available buffer space.  When failing, it doesn't unlock the write lock
of the ring buffer.
4. The next rx_try then fails to alloc, and doesn't call nrf_drv_uart_rx
to re-enable the receive interrupt.

The fix comes from the 15.3 version of the sdk.