Patches to stabilize uart
This commit contributes two patches to the Nordic SDK that seem to
stabilize the CLI over UART greatly.  It also adds a script I used to
stress test the UART (details for this are in comments in the script).

1. Patch fix-uart-error-handling.patch

Observed behavior: For some reason, after a while of typing manually in
the CLI, I inevitably get an overrun condition.  This means that the
6-bytes hardware rx fifo is already full when we receive another byte.
It is not clear to me why this happen, as I do not send a lot of data at
a high rate, but in any case we must be able to recover when this
happens.  After an overrun, the UART kept reporting an overrun forever.

Fix: This suggests to me that even when there's an overrun, the hardware
FIFO should be cleared.  I do this by reading the RXD register six
times.

Another observed behavior: In case of error, the ring buffer in
nrf_cli_uart was filled with nonsense.  The nrf_cli_uart event callback,
uart_event_handler, was called with a number of bytes
(p_event->data.error.rxtx.bytes) which was very improbable (255).  This
should contain data that was received and stored in the rx buffer (the
one passed to nrf_drv_uart_rx) before the error occured.  255 is the
typical buffer size we use (see the rx_try function).  In the irq
handler, this field seems to be set to wrong value.  It is set to
p_cb->rx_buffer_length, which is the total size of the rx buffer (as
passed to nrf_drv_uart_rx), not the size of data actually in that
buffer, p_cb->rx_counter.  That's the other change this patch does.

2. Patch nrf52/patches/fix-uart-cli-overrun-handling.patch

Observed behavior: If the software fifo, in nrf_cli_uart is full, when
we call rx_try, the UART rx irq will stay disabled and nothing will
enable it again, ever.  This leads to the rx just not working at this
point.

To kick things off at the start, cli_uart_enable calls rx_try.  This
allocates a chunk of bytes in the software rx fifo of the nrf_cli_uart
module and starts a uart rx operation by calling nrf_drv_uart_rx.  After
a timeout, or when the buffer is full, the nrf_drv_uart module uart
disables the receiving and calls nrf_cli_uart's handler
(uart_event_handler) with the data.  The handler finalizes the write to
the buffer (nrf_ringbuf_put), allocates a new chunk and starts a new rx
operation.  In the main loop, we try to call nrf_cli_process often
enough to consume from that software rx buffer.  If we are not able to
consume fast enough, there comes a point where rx_try tries to allocate
some space in the buffer, but there isn' any space left, so
nrf_ringbuf_alloc returns 0 bytes.  When this happens, we don't call
nrf_drv_uart_rx to kick off a new receive operation.  This breaks the
"chain" of receive operations, and nothing starts it again.

To fix this, I added a flag: when we try to allocate from the buffer but
are not able to, we raise the flag.  Later, when we consume from the
buffer (therefore creating room in the buffer), and notice this flag, we
start a new chain of rx operations.

With these two patches, I am able to spam the UART with hundreds of
times the "nsec\r\n" command, from a script, and the UART stays
functional.  Of course, we lose some data, since our system isn't able
to process all of that quick enough, but at least the UART stays alive.