Generate different external flash files per flavor
We will want different external flash contents per flavor.  Currently,
the external flash files (binary file and header) are not rebuilt when
rebuilding two flavors in a row like:

  $ make FLAVOR=ctf
  $ make FLAVOR=conf

Ideally, we would have a source directory and multiple, separate build
directories.  But we don't have that, so the next best thing I can think
of is to generate separate files per flavor.  This way, it's easier to
avoid including a header file generated for the wrong flavor.