Make setting icons smoother
I started from a large "gear" icon in gimp (~350x350px) and scaled it
down (using Cubic interpolation, which I think is the default).  This
gives some semi-transparent pixels around the edges of the shape, which
result in a much smoother result on the screen.

I am adding the GIMP source files, so that we can modify them easily if
needed.