cli: Make backspace and delete work by default
All terminal emulators I've tried send the character 0x7f (DEL) by
default, when pressing the backspace key.  This is currently interpreted
as the Nordic CLI as "delete the following character".  This patch makes
it delete the previous character instead, to match the expected behavior
of the backspace key.

All terminal emulators I've tried also send the sequence "\033[3~" when
pressing the delete button.  Handle it as well to produce the expected
result of deleting the following character.