Fix menu scrolling bugs, reduce number of unnecessary complete redraws
This patches makes us handle better the case where we are at the topmost
item and press up.  Before this patch, the result is that the last item
(now selected) is drawn at the top of the screen, the rest of the menu
being blank. After this patch, the last item (now selected) will drawn
at the bottom of the screen, with the preceding items visible.

It also fixes a bugs instances where we would redraw unnecessarily the
whole screen.  For example, when pressing up with the topmost item
selected or pressing down with the bottommost item selected, we don't
need to redraw everything if the menu entirely fits on one page.