Fix erroneous wrapping in text box
This fixes a bug where the last line of a text box page would be
invisible.

To reproduce at the previous commit, try to open the second talk of the
second conference day, titled "Cybersecurity vs the world".

The problem is that the first line is "Cybersecurity vs the world\n".
That is, 26 characters plus the trailing \n.  The text box is able to
show 26 characters.  When the "d" of "world" is printed, the gfx code
does some wrapping because the current line is full, it moves the cursor
to the second line.  We then print the trailing \n, which moves the
cursor to the third line.  We are now about to print the second line, at
the position of the third line.  Therefore, all subsequent lines are
shifted by one position below, and the last line falls off the screen.

Fix this by disabling wrapping while printing the text box.

The text wrapping is initialized at true and never set anywhere, so I am
a bit afraid that some other part depends on text wrapping but does not
explicitly enable it.  If we unilaterally disable it, we risk breaking
some other feature after the text box code executes.

So I made the text box code set the wrapping to off, but restore it to
its original value to leave it unchanged.