Fix choosing led colors multiple times in a row
If the user goes in the led color settings (e.g. Led first color), it
can change the color successfully once.  But if it tries again, the new
color won't apply.

The problem is the function call show_color_menu(0) at the end of
save_color, which unexpectedly changes _state to
SETTING_STATE_THIRD_COLOR.

Fix it by removing that call.