Improve instructions for flashing the badge
Update the instructions to add a bit more details.

I removed some mentions of using an external blackmagic, because I think
we should focus on the most probable scenario of a participant trying to
program his/her badge, who doesn't have an external blackmagic.

I added the instructions to flash the external flash, necessary to get a
fully working firmware.