Cast argument to isspace to unsigned char
I get this when compiling with gcc 9.2.0:

src/app/utils.c: In function 'word_wrap':
src/app/utils.c:80:29: error: array subscript has type 'char' [-Werror=char-subscripts]
   80 |         } else if (isspace(s[index])) {
      |                            ~^~~~~~~
cc1: all warnings being treated as errors

This is expected when passing a signed char to isspace.  The Linux man
page [1] for isspace says:

       The standards require that the argument c for these functions is
       either EOF or a value that is representable in the type unsigned
       char.  If the argument c is of type char, it must be cast to unsigned
       char, as in the following example:

           char c;
           ...
           res = toupper((unsigned char) c);

       This is necessary because char may be the equivalent of signed char,
       in which case a byte where the top bit is set would be sign extended
       when converting to int, yielding a value that is outside the range of
       unsigned char.

Add a cast to (unsigned char) to fix that.  The `volatile` is an earlier
attempt to get rid of this warning, but it doesn't do anything useful,
so remove it.

[1] http://man7.org/linux/man-pages/man3/isspace.3.html