Add home-made Makefile
Advantages over the current Makefile:

- We understand how it works and modify it to achieve what we want.
- Possibility to use different flags for our code, versus the SDK code
  (not done now, but it's on my todo list)
- It does completely* separate builds for each BOARD-FLAVOR combination.
  We could make it such that "make" in this directory builds all
  variants.

* Some files, like the source files generated from the images, are still
  generated in the source directory.  It's my goal to have them
  generated in the build directories later.

For now, it's a separate Makefile (Makefile.simark), which you can
invoke with:

  $ make -f Makefile.simark

This is so you can try both Makefiles side by side, compare, and find
any issue.  If this is deemed a good approach, then we can replace the
existing Makefile with this one.

Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>