makefile: manage header dependencies
One thing missing from the new Makefile is header dependency.  By that I
mean if you modify a header file, the next "make" should rebuild all the
source files directly or indirectly including this header.

The best practice way of doing this is to make the compiler output
Makefile rules that describe which headers are included by which source
files.  See the big comment for details.