makefile: replace nrf Makefile with the new one
Also adjust the top-level Makefile to the new names of the output files.

Tested by doing `make nrf -j 8` at the top-level.

Note that the output files must not exist in builds/nsec-badge-dev for
the builds to happen.  I would say that's an existing bug of the
top-level Makefile: it doesn't consider that things under nrf52/ might
have changed when you `make` again.