Add sdk patch to add CFI directives in app_error_handler
The function app_error_handler, provided by the SDK, is called whenever
an assertion fails (for example, APP_ERROR_CHECK fails).  This function
calls app_error_fault_handler, which is a user provided error handler
function.  app_error_handler is written in inline assembly, and modifies
the stack, which prevents GDB from unwinding the frame to look at caller
frames:

  (gdb) r
  The program being debugged has been started already.
  Start it from the beginning? (y or n) y
  Starting program: /home/smarchi/src/nsec/nsec-badge-dev/nrf52/builds/nsec19_nrf52_ctf.out

  Breakpoint 1, app_error_fault_handler (id=16385, pc=214639, info=536936408)
      at src/app/main.c:72
  72      void app_error_fault_handler(uint32_t id, uint32_t pc, uint32_t info) {
  (gdb) bt
  #0  app_error_fault_handler (id=16385, pc=214639, info=536936408) at src/app/main.c:72
  #1  0x0002a552 in app_error_handler (error_code=16385, line_num=214639,
      p_file_name=0x2000ffd8 "c")
      at nordicsdk/components/libraries/util/app_error_handler_gcc.c:49
  Backtrace stopped: previous frame identical to this frame (corrupt stack?)

It therefore prevents the user to know where the failure happened.  A
workaround is to break at the beginning of app_error_handler instead,
before it modifies the stack.

But really, app_error_handler should use CFI directives [1] to generate
debug information helping the debugger unwind that frame.  This patch
adds the minimal amount needed for the backtrace to work when we are
stopped inside app_error_fault_handler.  This is the result with the
patch applied:

  (gdb) r
  The program being debugged has been started already.
  Start it from the beginning? (y or n) y
  Starting program: /home/smarchi/src/nsec/nsec-badge-dev/nrf52/builds/nsec19_nrf52_ctf.out
  Note: automatically using hardware breakpoints for read-only addresses.

  Breakpoint 1, app_error_fault_handler (id=16385, pc=214639, info=536936408)
      at src/app/main.c:72
  72      void app_error_fault_handler(uint32_t id, uint32_t pc, uint32_t info) {
  (gdb) bt
  #0  app_error_fault_handler (id=16385, pc=214639, info=536936408) at src/app/main.c:72
  #1  0x0002a552 in app_error_handler (error_code=16385, line_num=214639, line_num@entry=99, p_file_name=0x2000ffd8 "c", p_file_name@entry=0x582fc "src/drivers/buttons.c")
      at nordicsdk/components/libraries/util/app_error_handler_gcc.c:49
  #2  0x0003466e in nsec_buttons_init () at src/drivers/buttons.c:99
  #3  0x0002fa58 in main () at src/app/main.c:177
  Prologue scan stopped at 0x262be

[1] https://sourceware.org/binutils/docs/as/CFI-directives.html