Move nordicsdk to nordicsdk-16.0.0
Following the SDK version bump, it is necessary to delete the nordicsdk
directory in order to download and extract the new SDK.  Some people
were confused because they still had an old SDK in their nordicsdk/, and
therefore failed to build the project.

By moving the SDK to a directory with the version name in it, we will
avoid this kind of clash.  The new SDK will automatically be downloaded.
The old nordicsdk directory will be left around, but it will be safe to
delete it.