Add a task for the CLI
Add a task that calls cli_process regularly to make the CLI work.  It
works, but is somehow very slow.  There's a delay between the keypresses
and the characters appearing on the screen, I'm not sure why.

Note that the CLI requires the app_timer module.  There is an
implementation of the app_timer module to use when also using FreeRTOS
(app_timer_freertos.c).  This in turn requires using the RTC as the tick
source for FreeRTOS, so switch to that (instead of systick).