Use appropriate implementation of APP_TIMER_TICKS
As of this commit, the CLI over UART works, but has a very noticeable
delay, each character takes more than one second to appear.  I traced it
down to the wrong implementation of the macro APP_TIMER_TICKS being
used.

The nrf_cli_uart module uses a 100 ms timer from the app_timer module to
manage receiving the bytes.  When FreeRTOS is used, an
alternative implementation of the app_timer module is used
(app_timer_freertos.c), which implements the app_timer API using the
FreeRTOS timers.

The app_timer.h file provides two implementations of the macro
APP_TIMER_TICKS (which converts a time in ms into a number of system
ticks), one for when using the standard app_timer and one for when using
the FreeRTOS version of app_timer.  We are currently using the
"standard" version of the macro, which gives a wrong number of ticks.
Instead of 100ms, we therefore get a number of ticks that corresponds to
roughly 1.6 seconds.

This patch sets the FREERTOS macro, which makes app_timer.h use the
right macro.  With this, the CLI feels as fast as usual (which means,
not very fast, but not too bad).