Enable NeoPixels
Initialize the code that controls the ws2812/NeoPixels.

For testing, I've made the "main" force some effect settings.  Of
course, it will have to be removed eventually.

It works (the LEDs light up), but it seems to me like the pattern is
going a bit fast, there might be some timing problems.