Compute run time stats
FreeRTOS is able to evaluate how much processor time each task uses.  We
need to provide it with a reasonnably high frequency timer (compared to
the tick frequency) that it can use to know how much time is spent on
each task.

This patch uses RTC2 for that, as RTC0 is already used by the
SoftDevice (trying to enable it causes a fault) and RTC1 is used to
generate the tick.

The run time stats are printed in the output of the CLI command
`freertos`.