ws2812fx: remove RESET_RUNTIME before allocation
The call to RESET_RUNTIME (which uses the `fx` variable) in
init_WS2812FX happens before the `fx` variable is initialized.  It
causes us to try to write at some unexpected place in memory.  Clearly,
it does not cause a crash, but it's not useful either, so remove it.
Note that there is another call to RESET_RUNTIME lower in the function.