Fix generation of gosecure animation sequences
When I introduced the new Makefile, I missed the generation of the
gosecure animation files.

The difficulty with these is that the list of these files that are
eventually included in the external flash is not know before-hand, so we
can't write proper make rules for them.  The generate.py script spews
out a bunch of .png files, which need to go in the external flash.

I did the best I could, which is explained in the Makefile, it's not
pretty but it works.