Add queue for button events
This commit adds a queue for button events.  When a button is pressed,
the nsec_button_event_handler function is called, through the magic of
the app_button and gpiote modules.

This function used to call nsec_controls_add_event, which would push a
button event in an nrf_queue.  Instead, this commit uses a FreeRTOS
queue.  This allows another task to wait on the queue and block as long
as there is no event.

I have updated the "UI" task to show this.