Start building the interface based on FreeRTOS
Home menu, main menu and settings menu are partially working.

The design looks a bit simpler than before.  Since there is a dedicated
UI RTOS task, we can block in the home_menu_application function,
waiting for user input, without affecting the rest of the system.  When
the user enters a submenu, we can call its function (e.g.
show_main_menu).  This function can in turn block on user input and
drive the submenu.  When it's done, its function returns and we
naturally come back to the home menu, in the same state we were.  The
home menu is however responsible to re-draw itself, knowing that the
submenu might have overwritten any part of the screen.