Remove global menu variable
With the new UI based on FreeRTOS, we'll want to have multiple menu
structures initialized at the same time.

Let's say you are in one menu you will be using a first menu structure.
If you enter a sub menu, we will initialize and use a second menu
structure.  When you press back, we'll come back to using the first menu
structure, in the same state that we have left it.

For this reason, we don't want to have a single global menu variable.
Each page that uses a menu will have its own.  For now, it is a
global static variable per file, but eventually it could be allocated on the
stack.

All menu functions now take a pointer to the menu as a parameter.  The
function menu_handler_init has been removed, as it is no longer needed:
the input / buttons won't work using this with FreeRTOS.