Add ui_page system
I started to find that I was copy pasting a lot of code in each page,
mainly the event loop that waits on button events.  This commit
introduces a concept of "ui_page" and a page stack.

Calling the show_ui_page function pushes a new page on the stack.  That
page will be able to draw itself and will now be the one receiving
button events.  When its button event handler returns true, it means it
wants to quit.  For example, when we are in a menu page and the user
presses back, that button handler would return true in order to quit the
page and go back to the previous one.

A page is a simple structure implementing three callbacks:

- init: called once when the page is entered
- draw: called once when the page is entered and when the controls comes
  back to this page, after a child page has quit
- button_handler: called when a button event is fired

By default, the execution blocks indefinitely on the button event queue.
However, if the page wants to do some periodic processing (for example,
the home page redrawing the animated logo), it can set its
ticks_timeout.  This is the amount of tick after which to give up
waiting for events, if none comes up.  The button handler will be called
with a new BUTTON_NONE event, allowing the page to do its processing.

Note that the page should not rely on the frequency at which its button
handler is called to keep a sense of time.  Even if the page uses a
tick_timeout, when the user presses buttons, the button handler will get
called before that tick timeout is expired.  It should instead use a
timer, or something like xTaskGetTickCount.