Fix race condition when setting neopixel settings
There is currently a possibility that changing the led colors through
the menu causes the UI to freeze.

This is because we are calling setArrayColor_packed_WS2812FX, which
eventually calls, setBrightness_WS2812FX, which calls
nsec_neoPixel_show, and then show_with_PWM.  While this happens, the
Neopixels task keeps calling service_WS2812FX, which also ends up
calling show_with_PWM.  Having two tasks configuring and using the PWM
device at the same time causes the UI task to block indefinitely waiting
for the PWM "end" event.

Only one task should be directly using the PWM.  Therefore, remove the
call to nsec_neoPixel_show from setBrightness_WS2812FX, such that
choosing a menu entry does not directly calls into the PWM code.
Choosing the menu entry will only set the led settings in the in-memory
structure (g_fx), which will be picked up the next time the Neopixels
task calls show_with_PWM.  This means there is a slight delay we feel
when changing a led setting, especially visible when the led speed is
set to the lowest.  This could be fixed by notifying the Neopixels task
when a setting changes, such that it would immediatly write those new
settings to the leds.  However, I left that as a future exercise, as
it will not be trivial.

Also, because the Neopixel reads the g_fx structure, and the UI task
could update it at the same time, the access to that structure needs to
be synchronized.  I therefore added a mutex that is taken whenever
calling into the ws2812fx code.