Specify nsec-badge-dev-esp32-external-components submodule address using relative path
When I try to `git clone --recursive`, it asks me my Github
username/password when cloning the
nsec-badge-dev-esp32-external-components submodule.  This is because
it's a private repo, and the address is an https one in .gitmodules.
Change it to be relative to the top-level repo instead.

This makes git use the same method to clone the submodule as it used for
the main repo, so everyone is happy.