Put all library JPEG images into one file
open() is an extremely slow operation on SPIFFS that can bring the
rendering to a complete halt, because of the expired filesystem cache
pages.

This operation can be completely avoided by putting all JPEG images into
a single file and manually managing read offsets and lengths.  With this,
JPEG operations take the same amount of time on average (because most of
the time is spent actually decoding the image,)  but the worst case slow
read can be avoided (which is very noticeable because of the dropped
framerate.)