Drop the ui_page component
The ui_page functionality was originally designed as a glue between a
legacy menu system of the badge and a more modular "application"
structure.

Since the old menu system is not migrated to the ESP32 firmware, the
role of the ui_page is reduced to simply receiving button events from
the events queue and passing them to the handler function.  Well, it
turns out that the handler function might as well do it directly: in a
loop, in a separate task, or not do it at all.  So at the moment, the
ui_page abstraction does not actually provide any useful functionality,
only limits the design choices for individual "applications."