Remove infinite loop at end of app_main
We currently have an infinite loop at the end of app_main, which is not
necessary.  If we let it finish, it goes back to the main_task function,
in file esp-idf/components/freertos/port/port_common.c.  It then does
`vTaskDelete(NULL)`, meaning the task deletes itself.  The remaining
tasks we have created keep doing their own thing.

This does not have an effect today though, because the code is stuck in
another infinite loop, in rpg_control_take.