Move extern "C" to header files
If we have C header files that we want to use from C++ source files,
it's more typical to put the extern "C" in the header files.  This
avoids having to repeat it everywhere.

I only added the extern "C" to graphics.h, because it's the only one
needed right now (in other words, it links fine like this).