Prevent NPC from accidentally jumping off scene
NPC movement is not uniform, it depends on the rendering speed.  It is
possible for a fast moving character to do a large leap forward and jump
over a small blocked section of the map.  Make all area outside of the
map area blocked by default to prevent them jumping off.