Initial commit for Conference Badge
Specs
- 3xAAA, atmega328, neopixels, input-port, output-port, optional oled

Story:
All conference goers will get a badge. They will need to go to the soldering village to solder their OLED screen and their buttons. The buttons are used to enter their custom names which will be shown on the OLED screen.
There is an "input" and "output" port on the left and right side of the badge. Those port is used to connect a badge with another. When a participants does so, the badge will play an 30 seconds loading animation using the neopixels. This will encourage the participants to interact with the other user during that delay. Once the animation is complete, +1 point will be added to the user's score. Briefly after the user disconnects the badge, their score will be represented as a progress bar using the neopixels. And if the OLED screen is attached, the actual numerical score will be shown.