Add badge logic skeleton port
Add a compile-tested skeleton port of the 2023 conference badge. This
port only includes the communication code, basic scheduling facilities,
button watcher, and utils on which the code depends.

The serial communication is entirely stubbed-out to mock-up the Arduino
interfaces.

Follow-up commits will setup the tasks to use the FreeRTOS scheduler and
wire-up the communication handler to the SDK's serial interfaces.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>