Exclude all components from "conf" firmware build
The conference and CTF firmwares will be separate. These changes to the
CMakeLists ensure only relevant components are shared. As of this
commit, the conference firmware is bare. Follow-up changes gradually
introduce more functionalities.

As of this commit, I can't change the build configuration to build CTF
or ADDON from menuconfig. The sdkconfig.conf file changes to reflect the
target change, but it seems to be ignored by CMake? (maybe the build
configuration isn't regenerated after a config change?)

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>