badge-network: use the esp-idf interfaces to setup communication
A `uart_interface` class is added to wrap the ESP-idf uart SDK.

It makes use of some additional scaffolding included as part of this
change, namely a basic logging class based around libfmt and a number of
exception handling helpers.

It seems that variables that are declared as part of `globals.cpp` are
ignored. I am not sure why, but I am guessing an over-eager
linker (or perhaps an option like no-as-needed is being passed).

I have not investigated that behaviour yet. For the moment, the
declaration of `badge` is moved to the entry point's translation unit
which seems to work around the problem since it prevented the
initialization of the badge's network manager.

The setup of the network manager is also performed as part of its
constructor since the constructor launches the task. We must ensure the
setup is fully performed by the time the task runs. It is also
preferable to perform that setup on the main thread to catch and handle
setup errors early.

As of this change, the wire protocol is operating in "idle"
mode (exchanging monitor messages back and forth) at 1Hz. The frequency
will be increased later, but for the moment it makes any debugging way
easier.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>