Move button callback registration to the button watcher
Move the button handling logic to the watcher. This allows the re-use of
existing C++ utilities and configuration files. It also completely
eliminates the need for a task.

The button SDK already makes use of the timers in task mode[1] which
makes it safe to invoke non-trivial callbacks from the button handling
code.

The watcher listens for BUTTON_PRESS_DOWN instead of the equivalent 'UP'
event. Overall, it gives a snappier feel when cycling the animations
since the cycling occurs on the press down rather than the release of
the button.

https://github.com/espressif/esp-iot-solution/blob/b2c33a0f72c0d4b968a72a3da0de8d8e3d29db83/components/button/iot_button.c#L356

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>