Enable console on USB
Unfortunatly the 'console' module provided by the ESP-IDF framework as
well as the logging system are tied to the default console configured in
sdkconfig. This means I won't be able to send the logging to UART0 and
have the console on the USB port without re-implementing the 'console'
module which I don't plan to do.

We will have to disable logging in the production firmware so it doesn't
interfere with the console.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>