Fix: buttons sometimes not properly initialized
It seems that FreeRTOS and/or the ESP-IDF SDK are not completely
initialized before entering app_main(). Since the badge object is
global, it's constructor runs before the control is handed over to
app_main().

The badge instance is moved within the scope of app_main() which causes
the initialization to happen within a proper FreeRTOS task (and not the
idle task, if I understand the FreeRTOS code correctly).

The periodic tasks are launched explicitly by calling badge::start()
since some of the code relies on a global `the_badge` pointer being
initialized. Starting the tasks as part of the badge's constructor would
race with the publication of that pointer.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>