badge: restore social display in "idle" state"
Restore the social level display while in "idle" state" that was removed bu commit "5ab1799972a".

With this commit, the social level and the health bar are alternating when the down button is pressed while "idle".

This commit keep the new implementation that display the health bar (or the social level) until the 'left" or "right" button is press.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>