Add a round-robin consumer class
Implemented a template-based consumer class for round-robin consumption
from multiple arrays, ensuring sequential access to elements from a
fixed number of arrays as long as any remain. The class skips exhausted
arrays and continues with the rest, utilizing a consume method for
element retrieval.

This is useful to implement the animation selection logic as we want to
round-robin between the various animation classes as the unlocked
animations increase (i.e. we don't want to sequentially unlock cycles,
shooting-star, and heartbeat animations).

This class is used in a follow-up commit.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>