Fix: badge: underflow when cycling animations
Copied from Sylvain Lemieux's commit message addressing the same
problem:

When the badge is running for the first time (i.e. social level 0) or is factory reset,
the selected animation is set to 255 (see log bellow).

[badge:info 0] Button event: button=LEFT, event=SINGLE_CLICK
[badge:info 0] Cycling selected animation: direction=-1, original_animation_id=0, new_animation_id=-1
[Strip animator:info 0] Idle animation set: id=255
[Config store:info 0] Saving field to namespace: namespace_name=badge2024, field_name=slctd_anim, value=255
[Config store:info 0] Saving field to namespace: namespace_name=badge2024, field_name=social_lvl, value=0
[Strip animator:info 0] Idle animation set: id=255

Add a verification to ensure the selected animation is set to 0 when the social level is 0.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>