strip_animator: Add support for 2 inactive colors in progress bar
Update the existing progress bar keyframe processing to support the
display of 2 inactive colors on the upper (0-7) and lower (8-15) LEDs.

This commit does not add any behavior changes to the current implementation.

This change is require to support a different color for the return
path (i.e. lower LEDs) during the badge synchronization process.

Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>