badge: bug fix - post sync animation health bar
During the post synchronization animation, when the health meter
is display, the social level used for the health meter mapping
is the old, instead of the new level.

Fix this issue by saving the new badge social level at the end
of the social level animation. When the health meter mapping is
calculated, the new social level is used.

Fixes: a5d670860f1f
Signed-off-by: Sylvain Lemieux <sylemieux.git@gmail.com>