Fix: badge: illegal double-enqueue of annoucement message
The left badge sometimes enters an error state from which it cannot
recover when it attempts to enqueue an 'ANNOUNCE_BADGE_ID' message while
an 'ANIMATION_DONE' message is already queued.

The network handler does not allow the other tasks to enqueue more than
one message as it would be difficult to reliably bound the number of
queued messages. Without a solid bound, it would have to resort to
dynamic allocations (which arguably would not be a problem on this class
of hardware, but was definitely problematic on the AVR-based badge of
last year).

As such, the application layer must pace itself and ensure it only
enqueues new messages after the successful tranmission of the last
enqueued message.

In this case, the problematic scenario on the left-most badge is:

The pairing animator completes the animation and enqueues
'ANIMATION_DONE' towards the right neighbor. It then transitions its
state to 'DONE'.

The 'DONE' state handler then immediately sets the network application
state to 'EXCHANGING_IDS'; that is the state during which the unique
badge IDs are actually transmitted.

As the network app state changes, the _id_exchanger's 'start()' method
is invoked, which immediately enqueues the badge's ID annoucement
message.

Depending on the timings, the network handler may or may not have sent
the ANIMATION_DONE message. If it is still in the outgoing queue, the
network_handler::enqueue_app_message() method will fail and log:
  Failed to enqueue message: outgoing queue is full [...]

To address this, network_id_exchanger::start() is modified to set
_send_ours_on_next_send_complete' to 'true'. This attribute is used by
the id_exchanger's new_message handler to determine if it should enqueue
a message announcing the badge's ID.

The fix is very short since it leverages the same logic that was already
used by the badges in the 'middle' link position.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@gmail.com>